// ClientApp/apps/vendor-management/src/Modules/Shared/Interfaces/Request/API-Request-BankInformation.interface.ts
export interface IAPIRequestBankInformation {
    userId: string;
    id: number;
    name: string;
    addressLine1: string,
    addressLine2: string,
    countryId: number;
    countryCode: string;
    countryName: string;
    stateId: number;
    stateCode: string;
    stateName: string;
    cityId: number;
    cityCode: string;
    cityName: string;
    zipCode: string;
    vendorId: number,
  }
  